+++ docs/release-notes.rst
@@ -80,6 +80,15 @@ Bug Fixes
   get assigned to the wrong entity.
 - Parallel execution (introduced in 0.8.0) had a bug in logging where log messages were
   dropped (with a warning) when any argument to the log message was unpickleable.
+- The cache API (introduced in 0.8.0) had a bug where if two cache entries point to the
+  same artifact, deleting one of them could leave the other in a bad state.
+
+Improvements
+............
+
+- Bionic now gracefully handles the situations where a cached artifact file is deleted
+  but the corresponding metadata entry is not. (It now deletes the invalid metadata file
+  and computes a new artifact and metadata entry.)
 
 0.8.1 (Jul 6, 2020)
 --------------------